+++ apps/api/src/app.module.ts
@@ -11,7 +11,7 @@ import { LoggerInterceptor } from './interceptors/logging.inspector';
 import { LoggerModule } from './logger/logger.module';
 import { PrefectureModule } from './prefecture/prefecture.module';
 import { PrefectureStatsMetadataModule } from './prefecture-stats-metadata/prefecture-stats-metadata.module';
-import { GameModule } from './todofuken/game/game.module';
+import { GameModule } from './todoufuken/game/game.module';
 import { UserMiddleware } from './user/user.middleware';
 import { WeaponModule } from './weapon/weapon.module';
