+++ apps/api/src/prisma/prisma.service.ts
@@ -1,7 +1,7 @@
 import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
 import { PrismaClient } from 'database';
 import { paginate } from 'prisma-extension-pagination';
-import { GameDifficulty } from 'schema/dist/todofuken/game';
+import { GameDifficulty } from 'schema/dist/todoufuken/game';
 
 export type PrismaService = ReturnType<BasePrismaService['withExtensions']> & {
   pg(): ReturnType<BasePrismaService['pg']>;
