+++ apps/web/src/constants/cookie-names.ts
@@ -1,4 +1,4 @@
 export const COOKIE_NAMES = {
   // AUTH_TOKEN: 'react-oauth-token', // firebaseの認証を使うようにしたのでcookieが不要になった
-  CURRENT_TODOFUKEN_GAME_ID: 'todofuken-game-id',
+  CURRENT_todoufuken_GAME_ID: 'todoufuken-game-id',
 };
