+++ apps/web/src/pages/(todoufuken)/_/api/submit-game-turn-act.ts
@@ -1,6 +1,6 @@
 import { Game } from 'database';
-import { GameLogResponse } from 'schema/dist/todofuken/game/log';
-import { CreateGameLogInput } from 'schema/dist/todofuken/game/log';
+import { GameLogResponse } from 'schema/dist/todoufuken/game/log';
+import { CreateGameLogInput } from 'schema/dist/todoufuken/game/log';
 
 import { axios } from '@/lib/axios';
 import { useMutation, MutationConfig } from '@/lib/react-query';
@@ -22,7 +22,7 @@ type UseSubmitGameLogTurnActOptions = {
 export const useSubmitGameTurnAct = ({ config }: UseSubmitGameLogTurnActOptions = {}) => {
   return useMutation({
     // onSuccess: (res) => {
-    //   queryClient.invalidateQueries([QUERY_KEYS.TODOFUKEN_GAMES, res.gameId]);
+    //   queryClient.invalidateQueries([QUERY_KEYS.TODOUFUKEN_GAMES, res.gameId]);
     // },
     ...config,
     mutationFn: submitGameTurnAct,
