+++ apps/web/src/pages/(todoufuken)/_/components/game-mode-info.tsx
