+++ apps/web/src/pages/(todoufuken)/_/config/game.tsx
