+++ packages/schema/src/todoufuken/game/game-difficulty.config.ts
