+++ packages/schema/src/todoufuken/game/game.schema.ts
