+++ packages/schema/src/todoufuken/game/index.ts
